Add seccomp_sigsys_decode and seccomp_sigsys_set_return_value
This can decode a SIGSYS to a struct seccomp_data.  It shows the raw
(untranslated) values as opposed to the translated values.  For
example, socketcall will show up as socketcall.

Decoding can be used by itself to print useful error messages, and
decoding can be used in conjunction with
seccomp_sigsys_set_return_value to emulate trapped syscalls.

This supports x86_64, i386, and ARM if the syscall architecture
matches the libseccomp architecture.  It also supports i386 compat
syscalls with a 64-bit libseccomp.  (That is unlikely to happen
without either some kind of thunk library to load 32-bit code into a
64-bit process or explicit invocations of compat syscalls from
64-bit code.)  It should also work on x32, although I haven't tested
it.

(NB: ARM, in theory, supports syscalls with seven arguments.  The
only syscall like this is sys_syscall, which is (AFAICT) only exists
because it's awkward to implement syscall(2) on OABI without it.
seccomp does not, and probably never will, support filtering on the
seventh argument of sys_syscall, and there's little point for us to
have a nasty special case just to be able to decode it..)

Signed-off-by: Andy Lutomirski <luto@amacapital.net>